+++ storage.go
@@ -4,6 +4,7 @@ import (
 	"database/sql"
 	"fmt"
 	"os"
+	"time"
 )
 
 var (
@@ -17,6 +18,19 @@ var (
 
 var dbInfo = fmt.Sprintf("host=%s port=%s user=%s password=%s dbname=%s sslmode=%s", host, port, user, password, dbname, sslmode)
 
+func StartDB() {
+	time.Sleep(5 * time.Second)
+	if os.Getenv("CREATE_TABLE") == "yes" {
+
+		if os.Getenv("DB_SWITCH") == "on" {
+
+			if err := createTable(); err != nil {
+				fmt.Printf("err: %v\n", err)
+			}
+		}
+	}
+}
+
 func connectDB() (db *sql.DB) {
 	db, err := sql.Open("postgres", dbInfo)
 	if err != nil {
@@ -25,11 +39,10 @@ func connectDB() (db *sql.DB) {
 	return
 }
 
+// Созадет таблицу если она еще не создана
 func createTable() error {
-	db, err := sql.Open("postgres", dbInfo)
-	if err != nil {
-		fmt.Printf("DATABASE ERROR: %v\n", err)
-	}
+	db := connectDB()
+	defer db.Close()
 	if _, err := db.Exec(`CREATE TABLE passwords(id SERIAL PRIMARY KEY, username TEXT, chat_id INT, service TEXT, password TEXT);`); err != nil {
 		fmt.Printf("%v\n", "create Table")
 		return err
@@ -40,6 +53,8 @@ func createTable() error {
 // Получает первый пароль от переданного сервиса
 func Password(chat_id int, service, username string) (string, error) {
 	db := connectDB()
+	defer db.Close()
+
 	var service_name, pass string
 	fmt.Printf("service: %v\n", service)
 	fmt.Printf("chat_id: %v\n", chat_id)
@@ -56,9 +71,10 @@ func Password(chat_id int, service, username string) (string, error) {
 // Добавляет одну запись
 func AddPassword(username string, chatid int64, service string, password string) error {
 	db := connectDB()
+	defer db.Close()
 	data := `INSERT INTO passwords(username, chat_id, service, password) VALUES($1, $2, $3, $4);`
-	if _, err := db.Exec(data, username, chatid, service, password); err != nil {
-		fmt.Printf("%v\n", "ADDPasssword")
+	if _, err := db.Exec(data, username, chatid, service, Encrypt(password)); err != nil {
+		fmt.Printf("Encrypt(password): %v\n", Encrypt(password))
 		return err
 	}
 	return nil
@@ -67,6 +83,7 @@ func AddPassword(username string, chatid int64, service string, password string)
 // Удаляет одну запись из таблицы
 func DelPassword(username string, chatid int64, service string) error {
 	db := connectDB()
+	defer db.Close()
 	data := `DELETE FROM passwords WHERE service = $1;`
 	if _, err := db.Exec(data, service); err != nil {
 		fmt.Printf("%v\n", "DELPasssword")
@@ -75,8 +92,10 @@ func DelPassword(username string, chatid int64, service string) error {
 	return nil
 }
 
+// Получает все пары из сервисов и паролей
 func UserData(chat_id int, username string) ([]RowServ, error) {
 	db := connectDB()
+	defer db.Close()
 	var service_name, pass string
 	var data []RowServ
 	rows, err := db.Query(`SELECT service, password FROM passwords WHERE chat_id = $1 AND username = $2;`, chat_id, username)
@@ -90,7 +109,8 @@ func UserData(chat_id int, username string) ([]RowServ, error) {
 		if err := rows.Scan(&service_name, &pass); err != nil {
 			return data, err
 		}
-		data = append(data, RowServ{service_name, pass})
+
+		data = append(data, RowServ{service_name, Decrypt(pass)})
 	}
 
 	if err = rows.Err(); err != nil {
