+++ types.go
@@ -1,0 +1,11 @@
+package main
+
+type Config struct {
+	TelegramBotToken string
+}
+
+// Хранит пару из сервиса и пароля к нему
+type RowServ struct {
+	service  string
+	password string
+}
